refactor: SimilarToRecentlyViewedArtworks section
